+++ crystals/__init__.py
@@ -5,7 +5,7 @@ This package allows for manipulation and modelling of atomic structures in cryst
 __author__ = "Laurent P. René de Cotret"
 __email__ = "laurent.renedecotret@mail.mcgill.ca"
 __license__ = "GPLv3"
-__version__ = "1.5.0"
+__version__ = "1.6.0"
 
 from .atom import Atom
 from .atom import Element